% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.E1}
\alias{.E1}
\title{Unbiased estimator}
\usage{
.E1(n, i, M, nonparametric, Q)
}
\arguments{
\item{n}{vector of sample size}

\item{i}{group index}

\item{M}{a matrix}
}
\description{
Unbiased estimator
}
\keyword{internal}
