% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.E2}
\alias{.E2}
\title{Unbiased estimator}
\usage{
.E2(n, i, M, nonparametric, Q)
}
\arguments{
\item{n}{vector of sample size}

\item{i}{group index}

\item{M}{a matrix}
}
\description{
Unbiased estimator
}
\keyword{internal}
