% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.hrm.sigcode}
\alias{.hrm.sigcode}
\title{Function for the output: significant p-values have on or more stars}
\usage{
.hrm.sigcode(value)
}
\arguments{
\item{value}{p-value}
}
\description{
Function for the output: significant p-values have on or more stars
}
\keyword{internal}
