% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DualEmpirical}
\alias{DualEmpirical}
\title{Function for the dual empirical matrix}
\usage{
DualEmpirical(Data, B)
}
\arguments{
\item{Data}{data.frame}

\item{B}{not used}
}
\description{
Function for the dual empirical matrix
}
\keyword{internal}
