% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DualEmpirical2}
\alias{DualEmpirical2}
\title{Function for the dual empirical matrix}
\usage{
DualEmpirical2(Data, B)
}
\arguments{
\item{Data}{data.frame}

\item{B}{part of the hypothesis matrix}
}
\description{
Function for the dual empirical matrix
}
\keyword{internal}
