% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repeated.R
\docType{data}
\name{EEG}
\alias{EEG}
\title{EEG data of 160 subjects}
\format{
A data frame with 6400 rows and 7 variables.
}
\usage{
data(EEG)
}
\description{
A dataset containing EEG data (Staffen et al., 2014) of 160 subjects, 4 variables are measured at ten different locations.
}
\details{
The columns are as follows:
\itemize{
  \item group. Diagnostic group of the subject: Alzheimer's Disease (AD), Mild Cognitive Impairment (MCI), Subject Cognitive Complaints (SCC+, SCC-).
  \item value. Measured data of a subject at a specific variable and region.
  \item sex. Sex of the subject: Male (M) or Female (W).
  \item subject. A unique identification of a subject.
  \item variable. The variales measured are activity, complexity, mobility and brain rate coded from 1 to 4.
  \item region. Frontal left/right, central left/right, temporal left/right, occipital left/right, parietal left/right coded as 1 to 10.
  \item dimension. Mixing variable and region together, levels range from 1 to 40.
}
}
\keyword{datasets}
