% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{gui.results}
\alias{gui.results}
\title{Function for displaying and saving the results of the function 'hrm_test' within the function 'hrm.GUI'}
\usage{
gui.results(result, factors, dec, sep)
}
\arguments{
\item{result}{data.frame from the ouput of the function hrm_test}

\item{factors}{list containing the column names of the factors, first elements are the wholeplot factors, second the subplot factors}

\item{dec}{decimal mark}

\item{sep}{seperation mark for saving}
}
\description{
Function for displaying and saving the results of the function 'hrm_test' within the function 'hrm.GUI'
}
\keyword{internal}
