% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{GUI_error}
\alias{GUI_error}
\title{Function for presenting error messages to the user}
\usage{
GUI_error(e, msg)
}
\arguments{
\item{e}{exception}

\item{msg}{message to be presented to the end user}
}
\description{
Function for presenting error messages to the user
}
\keyword{internal}
