% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4_sub4.R
\name{hrm.0w.4s}
\alias{hrm.0w.4s}
\title{Test for interaction of four subplot factors}
\usage{
hrm.0w.4s(
  X,
  alpha,
  factor1,
  factor2,
  factor3,
  factor4,
  subject,
  data,
  H = 1,
  text = "",
  nonparametric,
  ranked,
  varQGlobal,
  np.correction
)
}
\arguments{
\item{X}{dataframe containing the data in the long table format}

\item{alpha}{alpha level used for the test}

\item{factor1}{column name of the data frame X of the first factor variable}

\item{subject}{column name of the data frame X identifying the subjects}

\item{data}{column name of the response variable}

\item{H}{string specifying the hypothesis}

\item{text}{a string, which will be printed in the output}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for interaction of four subplot factors
}
\keyword{internal}
