% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1_sub0.R
\name{hrm.1w.0f}
\alias{hrm.1w.0f}
\title{Test for interaction of factor A and B}
\usage{
hrm.1w.0f(
  X,
  alpha,
  group,
  subject,
  data,
  H,
  text,
  nonparametric,
  ranked,
  varQGlobal
)
}
\arguments{
\item{X}{dataframe containing the data in the long table format}

\item{alpha}{alpha level used for the test}

\item{group}{column name of the data frame X specifying the groups}

\item{subject}{column name of the data frame X identifying the subjects}

\item{data}{column name of the response variable}

\item{H}{string specifying the hypothesis}

\item{text}{a string, which will be printed in the output}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for interaction of factor A and B
}
\keyword{internal}
