% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3_AC.R
\name{hrm.1w.2f.AC}
\alias{hrm.1w.2f.AC}
\title{Test for interaction of factor A and C}
\usage{
hrm.1w.2f.AC(X, alpha, group, factor1, factor2, subject, data)
}
\arguments{
\item{X}{list containing the data matrices of all groups}

\item{alpha}{alpha level used for the test}

\item{group}{column name of the data frame X specifying the groups}

\item{factor1}{column name of the data frame X of the first factor variable}

\item{factor2}{column name of the data frame X of the second factor variable}

\item{subject}{column name of the data frame X identifying the subjects}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for interaction of factor A and C
}
\keyword{internal}
