% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4_sub3.R
\name{hrm.1w.3f}
\alias{hrm.1w.3f}
\title{Test for 1 wholeplot and 3 subplot-factors}
\usage{
hrm.1w.3f(
  X,
  alpha,
  group,
  factor1,
  factor2,
  factor3,
  subject,
  data,
  S,
  K1,
  K2,
  K3,
  hypothesis,
  nonparametric,
  ranked,
  varQGlobal,
  np.correction
)
}
\arguments{
\item{X}{dataframe containing the data in the long table format}

\item{alpha}{alpha level used for the test}

\item{group}{column name of the data frame X specifying the groups}

\item{factor1}{column name of the data frame X of the first factor variable}

\item{factor2}{column name of the data frame X of the second factor variable crossed with the first one}

\item{factor3}{column name of the data frame X of the third factor variable crossed with the first one}

\item{subject}{column name of the data frame X identifying the subjects}

\item{data}{column name of the data frame X containing the observed values}

\item{S}{Matrix for the wholeplot factor}

\item{K1}{Matrix for the first subplot factor}

\item{K2}{Matrix for the second subplot factor}

\item{K3}{Matrix for the third subplot factor}

\item{hypothesis}{String which is printed in the console as ouput to indicate which hypothesis is tested.}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for 1 wholeplot and 3 subplot-factors
}
\keyword{internal}
