% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4_AC.R
\name{hrm.2w.2f.AC}
\alias{hrm.2w.2f.AC}
\title{Test for influence of factor A}
\usage{
hrm.2w.2f.AC(X, alpha, group, subgroup, factor1, factor2, subject, data)
}
\arguments{
\item{X}{list containing the data matrices of all groups}

\item{alpha}{alpha level used for the test}

\item{group}{column name of the data frame X specifying the groups}

\item{subgroup}{column name of the subgroups (crossed with groups)}

\item{subject}{column name of the data frame X identifying the subjects}

\item{data}{column name of the data frame X containing the measurement data}

\item{factor}{column name of the data frame X of within-subject factor}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for influence of factor A
}
\keyword{internal}
