% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{hrm.GUI}
\alias{hrm.GUI}
\title{Graphical User Interface for Testing Multi-Factor High-Dimensional Repeated Measures}
\usage{
hrm.GUI()
}
\value{
The results can be saved as LaTeX Code or as plain text. Additionally a plot of the group profiles an be saved when using one whole- and one subplot factor.
}
\description{
Graphical User Interface (R Package RGtk2 needed) for the Function 'hrm_test': Test for main effects and interaction effects of one or two between-subject factors and one, two or three within-subject factors (at most four factors can be used).
}
\keyword{internal}
