% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVHRM.R
\name{hrm.mv.1w.1f}
\alias{hrm.mv.1w.1f}
\title{MVHRM test}
\usage{
hrm.mv.1w.1f(
  X,
  alpha,
  group,
  factor1,
  subject,
  data,
  variable,
  formula,
  nonparametric
)
}
\arguments{
\item{X}{dataframe containing the data in the long table format}

\item{alpha}{alpha level used for the test}

\item{group}{column name of the data frame X specifying the groups}

\item{factor1}{column name of the data frame X of the first factor variable}

\item{subject}{column name of the data frame X identifying the subjects}

\item{data}{column name of the response variable}

\item{H}{string specifying the hypothesis}

\item{text}{a string, which will be printed in the output}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
MVHRM test
}
\keyword{internal}
