\name{HRM-package}
\alias{HRM-package}
\alias{HRM}
\docType{package}
\title{
Inference on Low- and High-Dimensional Multi-Group Repeated Measures Designs with Unequal Covariance Matrices
}
\description{
Tests for main and simple treatment effects, time effects, as well as treatment by time interactions in possibly high-dimensional multi-group repeated measures designs. The groups are allowed to have different variance-covariance matrices. The parametric test assumes the observations to follow a multivariate normal distribution; the nonparametric version tests with regard to nonparametric relative effects based on pseudo-ranks.
}
\author{
Martin Happ \email{statistics@happ.co.at} (ORCID: \url{https://orcid.org/0000-0003-0009-2665}),
Solomon W. Harrar, Arne C. Bathke.

Maintainer: Martin Happ \email{statistics@happ.co.at}
}
\references{
Happ, M., Harrar, S. W. and Bathke, A. C. (2016). Inference for low- and high-dimensional multigroup repeated measures designs with unequal covariance matrices. \emph{Biometrical Journal}, \bold{58}(4), 810--830. \doi{10.1002/bimj.201500064}

Happ, M., Harrar, S. W. and Bathke, A. C. (2017). High-dimensional Repeated Measures. \emph{Journal of Statistical Theory and Practice}, \bold{11}(3), 468--477. \doi{10.1080/15598608.2017.1307792}

Happ, M., Harrar, S. W. and Bathke, A. C. (2018). HRM: An R Package for Analysing High-dimensional Multi-factor Repeated Measures. \emph{The R Journal}, \bold{10}(1), 534--548. \doi{10.32614/RJ-2018-032}

Staffen, W., Strobl, N., Zauner, H., Hoeller, Y., Dobesberger, J. and Trinka, E. (2014). Combining SPECT and EEG analysis for assessment of disorders with amnestic symptoms to enhance accuracy in early diagnostics. Poster A19 presented at the 11th Annual Meeting of the Austrian Society of Neurology, 26th--29th March 2014, Salzburg, Austria.
}
\seealso{
\code{\link{hrm_test}} for the main function of the package.
}
\keyword{package}
