% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1_sub0.R
\name{hrm.test.1.none}
\alias{hrm.test.1.none}
\title{Test for main group effect (weighted/unweighted)}
\usage{
hrm.test.1.none(X, alpha, group, subject, data, formula, nonparametric)
}
\arguments{
\item{X}{dataframe containing the data in the long table format}

\item{alpha}{alpha level used for the test}

\item{group}{column name of the data frame X specifying the groups}

\item{subject}{column name of the data frame X identifying the subjects}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for main group effect (weighted/unweighted)
}
\keyword{internal}
