% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{hrm.test.2.two}
\alias{hrm.test.2.two}
\title{Test for two subplot factors}
\usage{
hrm.test.2.two(
  X,
  alpha,
  factor1,
  factor2,
  subject,
  data,
  formula,
  testing = rep(1, 3),
  nonparametric,
  np.correction
)
}
\arguments{
\item{X}{dataframe containing the data in the long table format}

\item{alpha}{alpha level used for the test}

\item{factor1}{column name of the data frame X of the first factor variable}

\item{factor2}{column name of the data frame X of the second factor variable}

\item{subject}{column name of the data frame X identifying the subjects}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for two subplot factors
}
\keyword{internal}
