% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repeated.R
\name{hrm.test.2.within}
\alias{hrm.test.2.within}
\title{Test for no main effects and interactino effects of one between-subject factor and two crossed within-subject factors}
\usage{
hrm.test.2.within(
  X,
  alpha,
  group,
  factor1,
  factor2,
  subject,
  data,
  testing = rep(1, 7),
  formula,
  nonparametric,
  np.correction
)
}
\arguments{
\item{X}{dataframe containing the data in the long table format}

\item{alpha}{alpha level used for the test}

\item{group}{column name of the data frame X specifying the groups}

\item{factor1}{column name of the data frame X of the first factor variable}

\item{factor2}{column name of the data frame X of the second factor variable}

\item{subject}{column name of the data frame X identifying the subjects}

\item{data}{column name of the data frame X containing the measurement data}

\item{testing}{vector specifying which hypotheses should be tested}

\item{formula}{formula object from the user input}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for no main effects and interactino effects of one between-subject factor and two crossed within-subject factors
}
\keyword{internal}
