% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{hrm.test.dataframe-deprecated}
\alias{hrm.test.dataframe-deprecated}
\title{hrm.test.dataframe}
\usage{
hrm.test.dataframe(data, alpha = 0.05, group , subgroup,
factor1, factor2, factor3, subject, response)
}
\arguments{
\item{data}{A data.frame containing the data}

\item{alpha}{alpha level used for the test}

\item{subject}{column name within the data frame identifying the subjects}

\item{group}{column name within the data frame specifying the first whole-plot factor}

\item{subgroup}{column name within the data frame specifying the second whole-plot factor}

\item{factor1}{column name within the data frame specifying the first sub-plot factor}

\item{factor2}{column name within the data frame specifying the second sub-plot factor}

\item{factor3}{column name within the data frame specifying the third sub-plot factor}

\item{response}{column name within the data frame specifying the measurements}
}
\description{
The function listed below is deprecated and will be defunct in the near future. Please use 'hrm.test' instead.
}
\seealso{
\code{\link{HRM-deprecated}}
}
\keyword{internal}
