% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repeated.R
\name{hrm.test.dataframe}
\alias{hrm.test.dataframe}
\title{Test for main effects and interaction effects.}
\usage{
hrm.test.dataframe(data, alpha = 0.05, group, subgroup, factor1, factor2,
  factor3, subject, response)
}
\arguments{
\item{data}{A data.frame containing the data}

\item{alpha}{alpha level used for the test}

\item{group}{column name within the data frame data specifying the groups}

\item{subgroup}{column name within the data frame data specifying the subgroups (crossed with groups)}

\item{factor1}{column name within the data frame data specifying the first subplot-factor}

\item{factor2}{column name within the data frame data specifying the the second subplot-factor}

\item{factor3}{column name within the data frame data specifying the the third subplot-factor}

\item{subject}{column name within the data frame X identifying the subjects}

\item{response}{column name within the data frame X containing the response variable}
}
\value{
Returns an object from class HRM containing
\item{result}{A dat frame with the results from the hypotheses tests.}
\item{alpha}{The type-I error rate which was used.}
\item{subject}{The column name identifying the subjects.}
\item{factors}{A list containing the whole- and subplot factors.}
}
\description{
Test for main effects and interaction effects of one or two between-subject factors and one, two or three within-subject factors (at most four factors can be used).
}
\examples{
n=c(45, 22, 57, 36)
X_1 = mvrnorm(n = n[1], mu = rep(0,40), Sigma = diag(40))
X_2 = mvrnorm(n = n[2], mu = rep(0,40), Sigma = diag(40))
X_3 = mvrnorm(n = n[3], mu = rep(0,40), Sigma = diag(40))
X_4 = mvrnorm(n = n[4], mu = rep(0,40), Sigma = diag(40))

X_1 = c(t(as.matrix(X_1)))
X_1 = data.frame(group = "SCC+", value = X_1)

X_2 = c(t(as.matrix(X_2)))
X_2 = data.frame(group= "SCC-", value = X_2)


X_3 = c(t(as.matrix(X_3)))
X_3 = data.frame(group = "MCI", value = X_3)


X_4 = c(t(as.matrix(X_4)))
X_4 = data.frame(group = "AD", value = X_4)

X = rbind(X_1, X_2, X_3, X_4)
X$subject = gl(160, 40)
X$variable = as.factor(rep(gl(4, 10),160))
X$region = as.factor(rep(rep(1:10,4),160))
X$subgroup = as.factor(rep(gl(2, 40),80))
X$group = as.factor(X$group)

hrm.test.dataframe(data=X,alpha=0.05,group="group",factor1="variable",
factor2="region",subject="subject",response="value")
}
\keyword{export}
