% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{hrm.test.matrix-deprecated}
\alias{hrm.test.matrix-deprecated}
\title{hrm.test.matrix}
\usage{
hrm.test.matrix(data, alpha = 0.05)
}
\arguments{
\item{data}{A list of matrices}

\item{alpha}{alpha level used for the test}
}
\description{
The function listed below is deprecated and will be defunct in the near future. Please use 'hrm.test' instead.
}
\seealso{
\code{\link{HRM-deprecated}}
}
\keyword{internal}
