% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrmInstall.R
\name{hrmInstall}
\alias{hrmInstall}
\title{hrmInstall}
\usage{
hrmInstall(pckgs, force = F)
}
\arguments{
\item{pckgs}{vector of hrm package names to install}

\item{force}{TRUE or FALSE force install?}
}
\description{
install one or more hrm packages.
}
\examples{
\dontrun{
hrmInstall('hrm')
}
}
