% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{confint.HRM}
\alias{confint.HRM}
\title{Function to calculate confidence intervals}
\usage{
\method{confint}{HRM}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object from class 'HRM' returned from the function hrm_test}

\item{parm}{currently ignored; all possible confidence intervals are calculated}

\item{level}{confidence level (FWER) used for calculating the inverals}

\item{...}{Further arguments passed to 'hrm_test' will be ignored}
}
\value{
Returns a data.frame with mean and 1-alpha confidence interval for each factor combintation
}
\description{
Function to calculate simultaneous, asymptotic (1-alpha) confidence intervals for an object of class 'HRM'.
}
\examples{
# hrm.test with a data.frame using a 'formula' object

# using the EEG dataset
?EEG

## Not run:
# z <- hrm_test(value ~ group*region*variable, subject = "subject", data = EEG)
# confint(z)
## End(Not run)
}
\keyword{export}
