% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{hrm.0w.2s}
\alias{hrm.0w.2s}
\title{Test for interaction of factor A and B}
\usage{
hrm.0w.2s(
  X,
  alpha,
  factor1,
  factor2,
  subject,
  data,
  H = "B",
  text = "",
  nonparametric,
  ranked,
  varQGlobal,
  np.correction,
  tmpQ1g,
  tmpQ2g
)
}
\arguments{
\item{X}{dataframe containing the data in the long table format}

\item{alpha}{alpha level used for the test}

\item{factor1}{column name of the data frame X of the first factor variable}

\item{subject}{column name of the data frame X identifying the subjects}

\item{data}{column name of the response variable}

\item{H}{string specifying the hypothesis}

\item{text}{a string, which will be printed in the output}

\item{group}{column name of the data frame X specifying the groups}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for interaction of factor A and B
}
\keyword{internal}
