% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hrm.plot}
\alias{hrm.plot}
\title{Plots profiles of the groups.}
\usage{
hrm.plot(X, group, factor1, subject, data, xlab = "dimension",
  ylab = "means")
}
\arguments{
\item{X}{list containing the data matrices of all groups}

\item{group}{column name of the data frame X specifying the groups}

\item{factor1}{column name of the data frame X of the first factor variable}

\item{subject}{column name of the data frame X identifying the subjects}

\item{data}{column name of the data frame X identifying the measured data}

\item{xlab}{label of the x-axis of the plot}

\item{ylab}{label of the y-axis of the plot}
}
\value{
Plots profiles of the groups.
}
\description{
Plots profiles of the groups.
}
\examples{
data(EEG)
head(EEG)

# plots profiles according to groups with
# subplot factor called 'dimension' in data.frame
hrm.plot(EEG,"group", "dimension", "subject", "value")
}

