% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hrm.plot}
\alias{hrm.plot}
\title{Plots profiles of the groups.}
\usage{
hrm.plot(data, group, factor1, subject, response, xlab = "dimension",
  ylab = "means")
}
\arguments{
\item{data}{A data.frame containing the data}

\item{group}{column name within the data frame data specifying the groups}

\item{factor1}{column name within the data frame data specifying the first subplot-factor}

\item{subject}{column name within the data frame X identifying the subjects}

\item{response}{column name within the data frame X containing the response variable}

\item{xlab}{label of the x-axis of the plot}

\item{ylab}{label of the y-axis of the plot}
}
\value{
Plots profiles of the groups.
}
\description{
Plots profiles of the groups in case of one whole- and one subplot-factor.
}
\examples{
data(EEG)
head(EEG)

# plots profiles according to groups with
# subplot-factor called dimension
hrm.plot(EEG,"group", "dimension", "subject", "value")
}
\keyword{export}
