\name{HRM-package}
\alias{HRM}
\docType{package}
\title{
Inference on low- and high-dimensional multi-group reapeted-measures designs with unequal covariance matrices.
}
\description{
Tests for main and simple treatment effects, time effects, as well as treatment by time interactions in possibly high-dimensional multi-group repeated measures designs. The groups are allowed to have different variance-covariance matrices but the observations must follow a multivariate normal distribution.
}
\details{
\tabular{ll}{
Package: \tab HRM\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2018-10-16\cr
License: \tab GPL-2 | GPL-3 \cr
}
}
\author{
Maintainer: martin.happ@aon.at
}
\references{
Happ, M., Harrar, S. W. and Bathke, A. C. (2016), Inference for low- and high-dimensional multigroup repeated measures designs with unequal covariance matrices. Biom. J., 58: 810-830. doi: 10.1002/bimj.201500064

Happ, M., Harrar, S. W. and Bathke, A. C. (2017), High-dimensional Repeated Measures. Journal of Statistical Theory and Practice. doi: 10.1080/15598608.2017.1307792

Staffen, W., Strobl, N., Zauner, H., Hoeller, Y., Dobesberger, J. and Trinka, E. (2014). Combining SPECT and EEG analysis for assessment of disorders with amnestic symptoms to enhance accuracy in early diagnostics. Poster A19 Presented at the 11th Annual Meeting of the Austrian Society of Neurology. 26th-29th March 2014, Salzburg, Austria.
}
\keyword{ package }
