\name{HRM-package}
\alias{HRM}
\docType{package}
\title{
Inference on low- and high-dimensional multi-group reapeted-measures designs with unequal covariance matrices.
}
\description{
Tests for main and simple treatment effects, time effects, as well as treatment by time interactions in possibly high-dimensional multi-group repeated measures designs.
}
\details{
\tabular{ll}{
Package: \tab HRM\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-09-08\cr
License: \tab GPL-2 | GPL-3 \cr
}
}
\author{
Maintainer: martin.happ@aon.at
}
\references{
Happ et al. (2015), Inference for low- and high-dimensional multi-group repeated measures designs with unequal covariance matrices, Biometrical Journal (in Revision)
}
\keyword{ package }
