\name{HRM-package}
\alias{HRM}
\docType{package}
\title{
Inference on low- and high-dimensional multi-group reapeted-measures designs with unequal covariance matrices.
}
\description{
Tests for main and simple treatment effects, time effects, as well as treatment by time interactions in possibly high-dimensional multi-group repeated measures designs.
}
\details{
\tabular{ll}{
Package: \tab HRM\cr
Type: \tab Package\cr
Version: \tab 0.3.5\cr
Date: \tab 2015-05-03\cr
License: \tab GPL-2 | GPL-3 \cr
}
}
\author{
Maintainer: martin.happ@aon.at
}
\references{
Happ, M., Harrar, S. W. and Bathke, A. C. (2015), Inference for low- and high-dimensional multigroup repeated measures designs with unequal covariance matrices. Biom. J.. doi: 10.1002/bimj.201500064

Staffen, W., Strobl, N., Zauner, H., Holler, Y., Dobesberger, J. and Trinka, E. (2014). Combining SPECT and EEG analysis for assessment of disorders with amnestic symptoms to enhance accuracy in early diagnostics. Poster A19 Presented at the 11th Annual Meeting of the Austrian Society of Neurology. 26th-29th March 2014, Salzburg, Austria.
}
\keyword{ package }
