% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repeated.R
\name{hrm.test.matrices}
\alias{hrm.test.matrices}
\title{Test for no main treatment effect, no main time effect, no simple treatment effect and no interaction between treatment and time}
\usage{
hrm.test.matrices(data, alpha = 0.05)
}
\arguments{
\item{data}{A list containing the data matrices of all groups. The rows are the independent subjects, these observations are assumed to be multivariate normally distributed. The columsn of all matrices need to be in the same order.}

\item{alpha}{alpha level used for the test}
}
\value{
Returns a data frame consisting of the degrees of freedom, the test value, the critical value and the p-value
}
\description{
Test for no main treatment effect, no main time effect, no simple treatment effect and no interaction between treatment and time
}
\keyword{internal}
